+++ src/skyborn/spharm/src/lap.f90
@@ -20,7 +20,7 @@ subroutine lap(dataspec, dataspec_lap, nmdim, nt, rsphere)
     rsphere_inv_sq = 1.0 / (rsphere * rsphere)
 
     ! Calculate truncation from nmdim
-    ntrunc = int(-1.5 + 0.5 * sqrt(9.0 - 8.0 * (1.0 - real(nmdim))))
+    ntrunc = -1.5 + 0.5 * sqrt(9.0 - 8.0 * (1.0 - real(nmdim)))
 
     ! Main computation loops - optimized for vectorization
     !$OMP PARALLEL DO PRIVATE(nmstrt, m, n, nm, n_real, lap_factor) &
